+++ java_warehouse/src/main/java/uz/jarvis/components/group/dtos/CompositeComponentInfoDTO.java
@@ -9,6 +9,6 @@ public class CompositeComponentInfoDTO {
   private Long id;
   private String code;
   private String name;
-  private Long checkpoint;
+  private Integer checkpoint;
   private Double quantity;
 }
