+++ java_warehouse/src/main/java/uz/jarvis/composite/CompositeCreateDTO.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class CompositeCreateDTO {
   private String code;
   private String name;
-  private Long checkpoint;
+  private Integer checkpoint;
   private String unit;
   private String specs;
   private Integer status;
